Rename: presentation 계층에 리턴할 응답 메시지 바디에 맞게 네이밍 수정
