+++ src/main/java/com/project/doubleshop/member/infrastructure/token/TokenRoleManager.java
@@ -8,8 +8,8 @@ import org.springframework.stereotype.Component;
 import com.project.doubleshop.common.Status;
 import com.project.doubleshop.member.domain.Member;
 import com.project.doubleshop.member.application.MemberService;
-import com.project.doubleshop.web.member.dto.RequestRole;
-import com.project.doubleshop.web.member.dto.ResultRole;
+import com.project.doubleshop.member.presentation.request.RequestRole;
+import com.project.doubleshop.member.presentation.response.ResponseRole;
 
 import lombok.RequiredArgsConstructor;
 
@@ -26,7 +26,7 @@ public class TokenRoleManager {
 	@Value("${admin-key}")
 	private String adminKey;
 
-	public ResultRole manage(HttpServletRequest request, Long memberId, RequestRole requestRole) {
+	public ResponseRole manage(HttpServletRequest request, Long memberId, RequestRole requestRole) {
 		Member member = authMemberService.findById(memberId);
 		if (member.getStatus().equals(Status.ACTIVATED)) {
 			if (requestRole.getAdminKey().equals(adminKey)) {
@@ -35,7 +35,7 @@ public class TokenRoleManager {
 				simpleToken.addRole(requestRole.getRole());
 				simpleToken.resetExpiry(simpleTokenConfigurer.getExpirySeconds() * 99);
 				tokenService.updateSession(tokenKey, simpleToken);
-				return new ResultRole(simpleToken.getUserId(), requestRole.getRole());
+				return new ResponseRole(simpleToken.getUserId(), requestRole.getRole());
 			} else {
 				throw new IllegalArgumentException("Invalid key provided.");
 			}
