+++ src/main/java/com/project/doubleshop/member/presentation/MemberInfoApi.java
@@ -2,9 +2,11 @@ package com.project.doubleshop.member.presentation;
 
 import java.util.Map;
 
+import javax.servlet.http.HttpServletRequest;
+
 import org.springframework.http.MediaType;
-import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -16,8 +18,8 @@ import org.springframework.web.bind.annotation.RestController;
 import com.project.doubleshop.common.ApiResult;
 import com.project.doubleshop.member.application.MemberFacade;
 import com.project.doubleshop.member.infrastructure.token.SimpleAuthentication;
-import com.project.doubleshop.web.member.dto.MemberResult;
-import com.project.doubleshop.web.member.dto.MemberInfoRequest;
+import com.project.doubleshop.member.presentation.response.MemberInfoResponse;
+import com.project.doubleshop.member.presentation.request.MemberInfoRequest;
 
 import lombok.RequiredArgsConstructor;
 
@@ -28,12 +30,12 @@ public class MemberInfoApi {
 	private final MemberFacade memberFacade;
 
 	@GetMapping(value = "member/me", produces = MediaType.APPLICATION_JSON_VALUE)
-	public ApiResult<MemberResult> me(@AuthenticationPrincipal SimpleAuthentication authentication) {
+	public ApiResult<MemberInfoResponse> me(@AuthenticationPrincipal SimpleAuthentication authentication) {
 		return ApiResult.OK(memberFacade.find(authentication));
 	}
 
 	@PatchMapping(value = "member/{id}/profile")
-	public ApiResult<MemberResult> update(@PathVariable Long id, @RequestBody MemberInfoRequest requestBody) {
+	public ApiResult<MemberInfoResponse> update(@PathVariable Long id, @RequestBody MemberInfoRequest requestBody) {
 		return ApiResult.OK(memberFacade.update(id, requestBody));
 	}
 
@@ -50,5 +52,11 @@ public class MemberInfoApi {
 	@PostMapping("member/exists/email")
 	public ApiResult<Boolean> checkDuplicateEmail(@RequestBody Map<String, String> requestMap) {
 		return ApiResult.OK(memberFacade.isDuplicated(requestMap));
+	}
+
+	@DeleteMapping("member/{id}/log-out")
+	public ApiResult<Boolean> logOut(HttpServletRequest request) {
+		String tokenKey = request.getHeader("x-auth-token");
+		return ApiResult.OK(memberFacade.invalidSession(tokenKey));
 	}
 }
