+++ src/main/java/com/project/doubleshop/member/presentation/response/MemberJoinResponse.java
@@ -1,4 +1,4 @@
-package com.project.doubleshop.web.member.dto;
+package com.project.doubleshop.member.presentation.response;
 
 import java.time.LocalDateTime;
 
@@ -12,7 +12,7 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor(access = AccessLevel.PRIVATE)
 @AllArgsConstructor
-public class JoinResult {
+public class MemberJoinResponse {
 
 	private String userId;
 
@@ -24,7 +24,7 @@ public class JoinResult {
 
 	private LocalDateTime createTime;
 
-	public JoinResult(Member member) {
+	public MemberJoinResponse(Member member) {
 		this.userId = member.getUserId();
 		this.name = member.getName();
 		this.email = member.getEmail();
