+++ src/main/java/com/project/doubleshop/member/presentation/response/ResponseRole.java
@@ -1,4 +1,4 @@
-package com.project.doubleshop.web.member.dto;
+package com.project.doubleshop.member.presentation.response;
 
 import com.project.doubleshop.member.infrastructure.token.Role;
 
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor(access = AccessLevel.PACKAGE)
 @AllArgsConstructor
 @Getter
-public class ResultRole {
+public class ResponseRole {
 	private String userId;
 	private Role role;
 }
